supervisord is now *not* called in the app's environment
The reason is that for dokku python apps, virtualenv is used to setup a
new python and module paths. The new python may be a different version
than apt-get's supervisord (which has a hardcoded shebang to
/usr/bin/python) which will result in the error: `ImportError: cannot
import name MAXREPEAT`. Also, virtualenv's python module paths does not
include the files required for apt-get installed supervisord, so
execution will fail. Therefore, the runner script is modified to NOT
setup the app's environment when executing supervisord. Instead, each
command is modified to pass through `/exec` so that the correct
environment is set.